requirements.txt
streamlit==1.45.0
langchain-community==0.3.25
faiss-cpu==1.11.0
numpy==1.26.4
gspread
oauth2client
python-dotenv
requests
